third_party/WebKit/LayoutTests/fast/css/invalidation/detach-reattach-shadow-expected.txt

CONSOLE WARNING: /deep/ combinator is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
PASS successfullyParsed is true

TEST COMPLETE
PASS getComputedStyle(inner).backgroundColor is "rgba(0, 0, 0, 0)"
